requirements.txt
boto3==1.7.36
botocore==1.10.36
docutils==0.14
jmespath==0.9.3
pycrypto==2.6.1
python-dateutil==2.7.3
s3transfer==0.1.13
six==1.11.0
